requirement.txt
Django
Django-ninja
